订单列表  订单详情搞完 ，差支付宝付款 ，和生成订单流程

+++ app/src/main/java/com/heloo/android/osmapp/model/OrderPriceBO.java
@@ -16,6 +16,6 @@ public class OrderPriceBO {
         public String icon;
         public Integer number;
         public Double discountprice;
-        public Double integralprice;
+        public Integer integralprice;
     }
 }
